2025-04-interactive-apps-with-databricks/backend/requirements.txt
fastapi
uvicorn
python-dotenv
databricks-connect
pydantic
python-dotenv